feat: Refactor components and add i18n support
This commit introduces two major improvements to the JB SQUARE application: a complete component refactoring and full internationalization (i18n) support.

Key Changes:

1.  **Component Refactoring:**
    - All UI sections (`Navigation`, `HeroSection`, etc.) have been extracted from `App.tsx` into their own files under `src/components/`.
    - Shared logic for animations and types has been moved to dedicated files (`animations.ts`, `types.ts`).
    - A custom hook, `useNavItems`, was created in `src/hooks/` to centralize and reuse the navigation data structure, preventing code duplication.

2.  **Internationalization (i18n):**
    - Integrated `react-i18next` to provide full bilingual (Korean/English) support.
    - Added a functional language toggle button in the navigation.
    - All user-facing text is now managed through `public/locales/` JSON files.

3.  **Enhancements & Fixes:**
    - Replaced potentially unreliable Unsplash image URLs with stable `picsum.photos` placeholders to fix loading issues.
    - Enhanced the footer with a newsletter subscription form and social media links.
    - Corrected the application's root entry point by ensuring `index.html` is properly configured to load the React app, resolving the 'no root page' issue.
