+++ src/components/Navigation.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { ChevronDown, Search, Menu, X } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
+import { useNavItems } from '../hooks/useNavItems';
 
 const dropdownVariants = {
   hidden: { opacity: 0, y: -10 },
@@ -14,14 +15,7 @@ const Navigation = () => {
   const [isScrolled, setIsScrolled] = useState(false);
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
   const [hoveredMenu, setHoveredMenu] = useState<string | null>(null);
-
-  const navItems = [
-    { title: t('nav.about'), key: 'about', subItems: [ { name: t('nav.jbsquare'), href: '#jbsquare' }, { name: t('nav.biovalley'), href: '#biovalley' }, { name: t('nav.supportorgs'), href: '#supportorgs' } ] },
-    { title: t('nav.opportunities'), key: 'opportunities', subItems: [ { name: t('nav.investment'), href: '#investment' }, { name: t('nav.announcements'), href: '#announcements' }, { name: t('nav.bizsupport'), href: '#bizsupport' } ] },
-    { title: t('nav.information'), key: 'information', subItems: [ { name: t('nav.companies'), href: '#companies' }, { name: t('nav.patents'), href: '#patents' } ] },
-    { title: t('nav.news'), key: 'news', subItems: [ { name: t('nav.news'), href: '#news' }, { name: t('nav.community'), href: '#community' } ] },
-    { title: t('nav.support'), key: 'support', subItems: [ { name: t('nav.contact'), href: '#contact' }, { name: t('nav.newsletter'), href: '#newsletter' } ] }
-  ];
+  const navItems = useNavItems();
 
   useEffect(() => {
     const handleScroll = () => setIsScrolled(window.scrollY > 10);
